requirements.txt
pytesseract
Pillow
langchain-openai
langchain-core
python-dotenv
geopy 
overpy
streamlit
playwright